saphire-backend/requirements.txt
alpha-vantage==2.1.3
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
autopep8==1.4.4
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-celery==3.3.1
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.7
numpy==1.18.2
pandas==1.0.3
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
supervisor==4.1.1.dev0
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
